[FEATURE] Statistics module: Better link labels
The labels of links in the statistics module are simply the URL of
the link.

This patch checks wheter a link points to a fragment (content element)
and retrieves the name of the content element.

If the link points directly to a page the link label is determined by
retrieving the page and extracting the <title> tag. Retrieving the page
record and using the 'title' field was not implemented because there are
many extensions and/or configuration options which modify the generated
page title. By the method used also link labels of external pages can get
displayed properly.

Internal page title will be retrieved using getURL, so it will get
any title the user in FE sees.

It also get the filename if the link is a filelink

Added TSConfig for admin to flexibly set, whether the content title,
the linked words or both should be shown.

Resolves: #46453
Releases: 3.0
Change-Id: If0551fd8bbd6cb75880091f3f5b8a1b36e8ea54c
Reviewed-on: http://review.typo3.org/21312
Reviewed-by: Ivan Dharma Kartolo <ivan.kartolo@gmail.com>
Tested-by: Ivan Dharma Kartolo <ivan.kartolo@gmail.com>