[FEATURE] Improve privacy of newsletter jumpurl links
When a link in a newsletter (plaintext/html) gets encoded via the jumpurl
feature there is a parameter "*&rid=" in the URL which represents the UID
of the users to whom the email was sent.

When the user clicks the link his UID is registered which could be seen as
a treat to privacy. If the site operator wants to connect the subscribers
to the links they clicked this is required but not for simple statistics.

This patch introduces a TS-Config option "jumpurl_tracking_privacy" which
allows to increase privacy by not adding the "&rid=" parameter to the
generated jumpurls. This does not change the functionality in any way
except the UIDs of the people who clicked a link get not recorded in
sys_dmail_log.

Added option in configuration module.
Check for empty rid in checkjumpurl script

Resolves: #48727
Releases: 3.0
Change-Id: Id04aebb67a26b631654c58b1dd803999f31642ac
Reviewed-on: https://review.typo3.org/21365
Reviewed-by: Bernhard Kraft
Tested-by: Bernhard Kraft