[BUGFIX] Fix possible jumpurl/dmailerping security leak
A recent change in the path check for dmailerping.gif introduced a possibel security leak.
As the passed jumpurl was only checked if it contains "dmailerping.gif" and then generates
a valid juHash it could be possible to retrieve arbitrary files using a NULL byte exploit.

This gets fixed by checking the submitted jumpurl using the core method "isAllowedAbsPath()"
and additionally checking if the jumpurl is simply a HTTP/HTTPS URL.

Change-Id: I588c26b3fc5e6b4d28e9ced0c45cab7be7636e20
Resolves: #48889
Reviewed-on: https://review.typo3.org/22118
Reviewed-by: Ivan Dharma Kartolo
Tested-by: Ivan Dharma Kartolo