[FEATURE] Allow to disable jumpurl for single links
Usually if the jumpurl feature is enabled every link in a HTML or plain
text mail will get jumpurl encoded. When the privacy feature is used
unsubscribe/change settings links in a newsletter should not get jumpurl
encoded because this would remove the "&rid=" attribute and make them
unusable.

Resolves: #52333
Releases: 3.1, 4.0
Change-Id: I2225ba57f10b19994c60c2d43b50bfc26af722f3
Reviewed-on: https://review.typo3.org/24110
Tested-by: Bernhard Kraft
Reviewed-by: Herbert Sojnik
Tested-by: Herbert Sojnik
Reviewed-by: Bernhard Kraft