[BUGFIX] Reapply: Fix assignment in class.dmailer.php
Corrected author comment in head of DirectMailEngineTest.php
because of which the previous commit has been reverted.

Additionally a unit test has been created for the method being
changed. This is the first unit test of the direct_mail extension
but further tests should get added for every change AFAP.

Change-Id: I849f751753e3235b41867fd2d2599ef3626091b5
Resolves: #53173
Reviewed-on: https://review.typo3.org/26377
Reviewed-by: Bernhard Kraft
Tested-by: Bernhard Kraft
Reviewed-by: Stefan Neufeind
Reviewed-on: https://review.typo3.org/26458