[BUGFIX] Can't run CLI script
When running the CLI script for sending mails (not the scheduler)
an PHP error gets generated.

Reason for this is the refactoring of the TYPO3 framework. The
class "t3lib_cli" was renamed to "CommandLineController" and
therefore has no constructor named "t3lib_cli" anymore.

Call the constructor via "parent::__construct()" instead.

Additionally also a unit test for this issued was introduced.

Change-Id: I6b0c6fae908f155c56c8781c7f7ec0e113b91901
Resolves: #54461
Releases: 3.1, 4.0
Reviewed-on: https://review.typo3.org/26455
Reviewed-by: Henrik Ziegenhain
Reviewed-by: Bernhard Kraft
Tested-by: Bernhard Kraft