[FEATURE] Allow creating mails from access protected pages
For internal newsletters, sometimes it is required to send pages that are
not publicly visible. Since TYPO3 uses the pageNotFound handler when such
pages are called directly, it it not possible to fetch such pages.

The determineId-PreProcessing hook in the TypoScriptFrontendController
can be used to override this behaviour. Direct Mail needs to add the uid
of a group that has access to the page being sent as an additional
parameter when fetching the page and the hook implementation then adds
this usergroup temporarily to the (anonymous) user fetching the page.
As different newsletters can have different groups assigned, the
usergroup used when fetching the mail content needs to be configurable
through TSconfig (just like the HTTP auth user and password).

For security reasons, a random access token is stored in the system
registry when the URL to fetch is generated. When the page is fetched,
this token is compared and removed immediately.

Resolves: #66296
Releases: 4.0
Change-Id: If722d3afb1f8fbaefee6318acb8ca4565b282c70
Reviewed-on: https://review.typo3.org/38498
Reviewed-by: Ivan Dharma Kartolo <ivan.kartolo@gmail.com>
Tested-by: Ivan Dharma Kartolo <ivan.kartolo@gmail.com>