[TASK] Improve parsing of link responses
Use DOMDocument instead of manual splitting of HTML code for link
responses.

Furthermore, the "Image link:" prefix that is not accurate (tracking
is also available for non-image links) is removed and the last 40
digits of the link are displayed by default instead of 40 characters
from the 8th character. It is more likely that the URL can be
identified by its end than by the beginning.

Resolves: #70963
Releases: 4.0
Change-Id: I0979666ccf69acd9b9a88d5c70b25c3d3a459cc1
Reviewed-on: https://review.typo3.org/44235
Reviewed-by: Ivan Dharma Kartolo <ivan.kartolo@gmail.com>
Tested-by: Ivan Dharma Kartolo <ivan.kartolo@gmail.com>