[BUGFIX] Space-separated recipient list: Check for null before using preg_split
In a normal, space-separated recipient list, the addresses are stored in a MEDIUMBLOB field. This field is nullable by default, therefore if no list has been entered, the field will have a NULL value. This breaks displaying or using recipient lists, becase preg_split cannot be applied on a NULL value. This is prevented by adding a check before using preg_split.